requirements.txt
rich>=13.7.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
tavily-python>=0.5.0
